feat: email lazy loading
