+++ apps/Conduit-UI/src/pages/email/send.tsx
@@ -1,6 +1,13 @@
 import React, { ReactElement } from 'react';
 import EmailsLayout from '../../components/navigation/InnerLayouts/emailsLayout';
-import SendEmailForm from '../../components/emails/SendEmailForm';
+import dynamic from 'next/dynamic';
+import ScaleLoader from 'react-spinners/ScaleLoader';
+
+const SendEmailForm = dynamic(() => import('../../components/emails/SendEmailForm'), {
+  loading: () => (
+    <ScaleLoader speedMultiplier={3} color={'#07D9C4'} loading={true} height={21} width={4} />
+  ),
+});
 
 const Send = () => {
   return <SendEmailForm />;
